perf(web): cut streaming memory churn — memoized conversation body, SDK throttle, values trim
Part 1 (quick fixes) of #309. Streaming a long agent run grew the tab to
~1.5GB from quadratic allocation churn across the streaming pipeline.

Frontend:
- Extract the conversation render tree into a memoized ConversationBody
  fed only throttled (~16Hz) / identity-stable props, so the component
  owning useStream renders almost nothing per SSE event (previously the
  ~1,500-line ChatPage re-rendered per token). Pure message helpers move
  to message-helpers.ts.
- Subagent-only updates change no top-level value identity (the SDK only
  bumps an internal version), so a throttled useStreamTick prop keeps the
  memoized body following them at a bounded rate.
- Stop reading the SDK's thread.toolCalls getter (full message rescan per
  property read, per render); derive tool calls once per throttled
  messages change via computeToolCallsFromMessages.
- Pass throttle: true to useStream to coalesce same-tick SSE bursts
  (reattach replay is the worst case) into one notification per macrotask.
- Latch ModelUnavailableError / StaleInterruptError in onError instead of
  effects watching thread.error.
- Upgrade streamdown 1.6.11 → 2.6.0: 2.5 stops re-rendering completed
  blocks while streaming. Drop an accidental div-prop spread into
  Streamdown in reasoning.tsx that v2's stricter types reject, and track
  reasoning start time in a ref (never rendered).

Backend:
- Strip the deepagents `files` dict from streamed `values` snapshots
  (re-sent every superstep, can be megabytes on sandbox-heavy threads);
  no stream consumer reads it — the web client uses only messages/todos/
  __interrupt__, Slack ignores values, and history reads come from the
  checkpoint.

Refs #309 (Part 2 — @langchain/react + Agent Streaming Protocol
migration — remains open).
