+++ web/package.json
@@ -62,7 +62,7 @@
 		"shiki": "^3.15.0",
 		"snakecase-keys": "^9.0.2",
 		"sonner": "^2.0.7",
-		"streamdown": "^1.6.10",
+		"streamdown": "^2.6.0",
 		"tailwind-merge": "^3.4.0",
 		"tokenlens": "^1.3.1",
 		"use-stick-to-bottom": "^1.1.1",
