+++ web/src/components/ai-elements/reasoning.tsx
@@ -9,7 +9,14 @@ import {
 import { cn } from "@/lib/utils";
 import { ChevronDownIcon } from "lucide-react";
 import type { ComponentProps, ReactNode } from "react";
-import { createContext, memo, useContext, useEffect, useState } from "react";
+import {
+	createContext,
+	memo,
+	useContext,
+	useEffect,
+	useRef,
+	useState,
+} from "react";
 import { Streamdown } from "streamdown";
 import { Shimmer } from "./shimmer";
 import { Loader } from "@/components/ai-elements/loader";
@@ -64,19 +71,21 @@ export const Reasoning = memo(
 		});
 
 		const [hasAutoClosed, setHasAutoClosed] = useState(false);
-		const [startTime, setStartTime] = useState<number | null>(null);
+		// Never rendered — only read back when streaming ends to compute the
+		// duration, so a ref (not state) avoids a cascading render per stream.
+		const startTimeRef = useRef<number | null>(null);
 
 		// Track duration when streaming starts and ends
 		useEffect(() => {
 			if (isStreaming) {
-				if (startTime === null) {
-					setStartTime(Date.now());
+				if (startTimeRef.current === null) {
+					startTimeRef.current = Date.now();
 				}
-			} else if (startTime !== null) {
-				setDuration(Math.ceil((Date.now() - startTime) / MS_IN_S));
-				setStartTime(null);
+			} else if (startTimeRef.current !== null) {
+				setDuration(Math.ceil((Date.now() - startTimeRef.current) / MS_IN_S));
+				startTimeRef.current = null;
 			}
-		}, [isStreaming, startTime, setDuration]);
+		}, [isStreaming, setDuration]);
 
 		// Auto-open when streaming starts
 		useEffect(() => {
@@ -93,7 +102,9 @@ export const Reasoning = memo(
 					setHasAutoClosed(true);
 				}, AUTO_CLOSE_DELAY);
 
-				return () => clearTimeout(timer);
+				return () => {
+					clearTimeout(timer);
+				};
 			}
 		}, [isStreaming, isOpen, defaultOpen, setIsOpen, hasAutoClosed]);
 
@@ -202,7 +213,9 @@ export const ReasoningContent = memo(
 							className
 						)}
 					>
-						<Streamdown {...props}>{children}</Streamdown>
+						{/* Streamdown 2.x types props strictly (e.g. dir is a literal
+						    union) — the Collapsible div props never belonged on it. */}
+						<Streamdown>{children}</Streamdown>
 					</div>
 				</div>
 			</div>
